% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bench_date}
\alias{bench_date}
\title{A benchmark dataset with different date formats}
\format{
An object of class \code{character} of length 93.
}
\usage{
bench_date
}
\description{
A benchmark dataset with different date formats
}
\keyword{datasets}
