% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{chronos-package}
\alias{chronos-package}
\title{chronos: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
