% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{chronos}
\alias{chronos}
\title{General purpose parser for Dates from input data}
\usage{
chronos(
  x,
  formats = NULL,
  out_datetime = "\%Y-\%m-\%d \%H:\%M:\%S",
  out_date = "\%Y-\%m-\%d"
)
}
\arguments{
\item{x}{A vector of type character with date(time) expressions to be parsed and converted.}

\item{formats}{character vector of formats to try out (see \link[base:strptime]{base::strptime}).
If NULL, uses a set of predefined formats mostly taken from the anytime package.}

\item{out_datetime}{character defining the datetime format of the parsed strings}

\item{out_date}{character defining the date format of the parsed strings}
}
\value{
A character vector which can be transformed to \code{POSIXct} or date
}
\description{
This function uses the dateparser and chrono crates from Rust to parse datetimes (and
dates) from strings.
}
\examples{
chronos(bench_date)
}
\seealso{
\link{parse_datetime} and \link{parse_date} if you need more control over formatting
}
