% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{parse_date}
\alias{parse_date}
\title{Parse date from strings using different formats}
\usage{
parse_date(x, formats, out_format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{A vector of type character with date(time) expressions to be parsed and converted.}

\item{formats}{character vector of formats to try out (see \link[base:strptime]{base::strptime}).
If missing, uses a set of predefined formats from the anytime package.}

\item{out_format}{character defining the format of the parsed strings}
}
\value{
character vector of parsed dates.
}
\description{
Parse date from strings using different formats
}
