% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{parse_epoch}
\alias{parse_epoch}
\title{Parse datetime from epoch}
\usage{
parse_epoch(x, out_datetime = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{x}{A vector of type character with date(time) expressions to be parsed and converted.}
}
\value{
character vector of parsed dates.
}
\description{
Parse datetime from epoch
}
