% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{timeless-package}
\alias{timeless}
\alias{timeless-package}
\title{timeless: Fast General Purpose Date/Time Converter}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast general purpose date/time converter using Rust. The package wraps the crates 'dateparser' and 'chrono' and implements date time, date and epoch time parser for heterogeneous vectors of dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/timeless}
  \item \url{https://schochastics.github.io/timeless/}
  \item Report bugs at \url{https://github.com/schochastics/timeless/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Other contributors:
\itemize{
  \item Rollie Ma (author of dateparser : <https://github.com/waltzofpearls/dateparser>) [contributor, copyright holder]
  \item Brandon W. Maister (author of chrono : <https://github.com/chronotope/chrono>) [contributor, copyright holder]
  \item Dirkjan Ochtman (author of chrono : <https://github.com/chronotope/chrono>) [contributor, copyright holder]
  \item Seonghoon Kang (author of chrono : <https://github.com/chronotope/chrono>) [contributor, copyright holder]
  \item Eric Sheppard (author of chrono : <https://github.com/chronotope/chrono>) [contributor, copyright holder]
  \item Paul Dicker (author of chrono : <https://github.com/chronotope/chrono>) [contributor, copyright holder]
}

}
