% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{chronos-package}
\alias{chronos-package}
\title{chronos: General purpose date(time) parser}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

General purpose date(time) parser. The package wraps the Rust crate 'dateparser'.
}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

}
