% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{chronos-package}
\alias{chronos-package}
\title{chronos: Fast General Purpose Date/Time Converter}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast general purpose date/time converter. The package wraps the Rust crate 'dateparser'.
}
