% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{chronos}
\alias{chronos}
\title{Parse Date from input data}
\usage{
chronos(x)
}
\arguments{
\item{x}{A vector of type character with date(time) expressions to be parsed and converted.}
}
\value{
A character vector which can be transformed to \code{POSIXct} or date
}
\description{
This function uses the dateparser crate from Rust to parse datetimes (and
dates) from strings.
}
\examples{
chronos(bench_date)
}
