% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{parse_date}
\alias{parse_date}
\title{Parse date from strings using different formats}
\usage{
parse_date(x, formats = NULL, out_date = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{A vector with date(time) expressions to be parsed and converted.}

\item{formats}{character vector of formats to try out (see \link[base:strptime]{base::strptime}).
If NULL, uses a set of predefined formats mostly taken from the anytime package.}

\item{out_date}{character defining the date format of the parsed strings}
}
\value{
character vector of parsed dates.
}
\description{
Parse date from strings using different formats
}
