% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetime from strings using different formats}
\usage{
parse_datetime(x, formats = NULL, out_datetime = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{x}{A vector with date(time) expressions to be parsed and converted.}

\item{formats}{character vector of formats to try out (see \link[base:strptime]{base::strptime}).
If NULL, uses a set of predefined formats mostly taken from the anytime package.}

\item{out_datetime}{character defining the datetime format of the parsed strings}
}
\value{
character vector of parsed datetimes
}
\description{
Parse datetime from strings using different formats
}
